% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception_handling.R
\name{\%decorates\%}
\alias{\%decorates\%}
\title{Defines a generic decorator to be used by other closures (or functions within functions).
Reference http://weinstockj.github.io/blog/2016/06/08/decorators-in-r/ for more information.
Decorators are a concept which is borrowed from Python. In Python, they are built into the
language, but in R we must first define a Decorator operator just as this one. handle_exceptions()
is a function which is meant to decorate other functions.}
\usage{
decorator \%decorates\% f
}
\arguments{
\item{decorator}{a function to decorate another function}

\item{f}{a function which is decorated}
}
\description{
Defines a generic decorator to be used by other closures (or functions within functions).
Reference http://weinstockj.github.io/blog/2016/06/08/decorators-in-r/ for more information.
Decorators are a concept which is borrowed from Python. In Python, they are built into the
language, but in R we must first define a Decorator operator just as this one. handle_exceptions()
is a function which is meant to decorate other functions.
}
