% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception_handling.R
\name{handle_exceptions}
\alias{handle_exceptions}
\title{safely handle errors and warnings in any given function}
\usage{
handle_exceptions(expr)
}
\arguments{
\item{expr}{an expression or function}
}
\description{
safely handle errors and warnings in any given function
}
