% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_from_path.R
\name{read_data_from_path}
\alias{read_data_from_path}
\title{Read in any data from a file path provided by the user}
\usage{
read_data_from_path()
}
\value{
a tibble with the read in data
}
\description{
Read in any data from a file path provided by the user
}
