% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n_groups.R
\name{sample_n_groups}
\alias{sample_n_groups}
\title{dplyr::sample_n lacks the functionality of sampling group_by variables. Instead,
it is only capable of sampling rows in a table. sample_n_groups() samples a
number of groups in a previously grouped_by tibble. \strong{the tibble must be grouped
before calling the function}.}
\usage{
sample_n_groups(tbl, size, replace = FALSE, weight = NULL)
}
\arguments{
\item{tbl}{A tibble}

\item{size}{number of groups to sample}

\item{replace}{sample with or without replacement}

\item{weight}{sampling weights}
}
\value{
A tibble
}
\description{
dplyr::sample_n lacks the functionality of sampling group_by variables. Instead,
it is only capable of sampling rows in a table. sample_n_groups() samples a
number of groups in a previously grouped_by tibble. \strong{the tibble must be grouped
before calling the function}.
}
