% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_type_colnames_by_num.R
\name{split_type_colnames_by_num}
\alias{split_type_colnames_by_num}
\title{Splits all columns of same type into n groups (3 by default). Useful for
datasets with large numbers of variables and observations. Allows easier
exploration on heavier visualizations across groups of variables.}
\usage{
split_type_colnames_by_num(data, n = 3,
  data_type = .is_categorical_col)
}
\arguments{
\item{data}{a tibble}

\item{n}{number of groups to split by}
}
\value{
a vector of grouped column names of the same type
}
\description{
Splits all columns of same type into n groups (3 by default). Useful for
datasets with large numbers of variables and observations. Allows easier
exploration on heavier visualizations across groups of variables.
}
