% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception_handling.R
\name{handle_exceptions}
\alias{handle_exceptions}
\title{handle exceptions: decorator
safely handle errors and warnings in any given function by applying
decorator to function}
\usage{
handle_exceptions(f)
}
\arguments{
\item{expr}{an expression or function}
}
\description{
handle exceptions: decorator
safely handle errors and warnings in any given function by applying
decorator to function
}
