% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_public.R
\name{sample_n_groups}
\alias{sample_n_groups}
\title{Gather, group_by the generated key and value, count, then slice
to create a frequency table of the tibble. Functions with multiple
variables.}
\usage{
sample_n_groups(tbl, size, replace = FALSE, weight = NULL)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble
}
\description{
Gather, group_by the generated key and value, count, then slice
to create a frequency table of the tibble. Functions with multiple
variables.
}
